requirements_for_tests.txt
-r requirements.txt
PyHamcrest
nose
pep8